Implemented artemis_icm20649_read() which reads both accel and gyro raw data from the sensor.

+++ src/artemis_spi.h
@@ -19,7 +19,7 @@ extern "C" {
 
 typedef struct s_artemis_spi_t
 {
-    uint32_t chipselect;
+    uint8_t chipselect;
     artemis_iom_t iom;
 } artemis_spi_t;
